Add index updater (#703)
This speeds up the index by adding an Updater with a cache, responsible for updating
the index to the latest chain tip.

The Updater exploits the fact that many UTXOs are destroyed soon after they are
created. By maintaining a cache of outpoint to ordinal ranges in an in-memory
hashmap, we avoid a great deal of database inserts and removals. On mainnet
this avoids around 80% of database writes and results in a 2x speedup at the
expense of slightly higher RAM usage.